+++ CHANGES.rst
@@ -11,6 +11,9 @@ Changelog
 - Optimized `MeetingItem.updateLocalRoles`, pass `cfg` and `item_state` when
   possible and `ram.cache` for `utils.compute_item_roles_to_assign_to_suffixes`.
   [gbastien]
+- Removed `Meeting.items` `ReferenceField`, manage it manually,
+  this will help migrating to `DX`.
+  [gbastien]
 
 4.2b6 (2020-11-19)
 ------------------